Added revert to delete LPAR from destination host on failed rebuild
This change adds a revert method to the Create task. Originally
there was no revert task for this method because LPAR cleanup
was handled elsewhere on a failed build. On rebuilds however,
this cleanup code was not being run because new builds and
rebuilds follow different flows. This was causing future
rebuilds to fail, because it appeared that the LPAR was already
on the destination host. The added revert method checks to
confirm that the failed build was a rebuild, and if it was,
it removes the LPAR off the destination host.

Change-Id: I56bae7e545f56d1fdf455309d11f24b61b0a9a1c
Closes-Bug: #1596797