Deploy of a VM occasionally fails due to invalid keystone token
Deploy of virtual machine occasionally fails due to invalid/expired
keystone token with swift and this causes the NVRAM data upload
operation to fail with ClientException error. The results object
returned to nova_powervm will have the error information appended
to it by the swiftclient. Hence, we loop through the results
object to see if there's an authentication failure, and if there
is, we'll retry upload operation, this time with a valid keystone token.

Change-Id: I2779ad6340a58f5c58971eb142ec7b0bd76e2016
Closes-Bug: #1607772