Move LPAR scrub tasks from build_map to Create task
In response to https://review.openstack.org/#/c/324730/3 we need to
be able to create the build_map sooner in spawn flow, before the
LPAR is created. To do this we need to move the LPAR scrub out of
the build_map initialization and into the Create task right after
the LPAR is created.

Partial-Bug: #1587942
Change-Id: Ie3879836c593703b215c54c67abeb358ea0bda82