Support just waiting for created vif events
In a previous change, we introduced the concept of new VIFs versus
updated VIFs.  This change fixes an issue where the create flow would
wait for neutron to send events for ALL of the vifs.  It should only be
waiting for events for the created vifs, otherwise it waits until the 5
minute timeout and fails.

Change-Id: I02fe4d32e1d4e83672758a8d143153ecbb227515
Closes-Bug: 1618208