More Specific Error Messages on get-vnc-console
When the open_remotable_vnc_vterm method raises a VNCBasedTerminal
FailedToOpen exception this is turned into a handled ConsoleType
Unavailable exception, but if the method fails for any other reason
the exception thrown is something isn't handled which causes the
REST API to show an unknown error.

No matter what the reason for the vterm failing to open, it would
be useful to raise a handle exception with the real exception string
embedded with in it. Also, in cases such as when the LPAR does not
exist, it would be useful to give a more specific error rather than
the long HTTP response error information from the 404.

This changeset will change it so an exception occuring on the
open_remotable_vnc_vterm method will be turned into a ConsoleType
Unavailable exception with the embedded error string in it.  Also,
it will check specifically for a 404 from the HTTP response looking
up the LPAR, and put out a specific error message saying it doesn't
exist.

Change-Id: Icdb2592dbcc520d4d0a7f4556a8b5877a4f31979
Closes-Bug: #1619153