HttpError=>VirtualInterface(Unp|P)lugException
With this change set, the top-level vif driver plug/unplug methods trap
HttpError (typically a 400 or 500) and convert it to the standard
VirtualInterfacePlugException/VirtualInterfaceUnplugException from
nova.exception.

This change set also gets rid of the VirtualInterfaceUnplugException in
vif.py and tasks/network.py, replacing their invocations with the
corresponding nova.exception class.

Conflicts:
	nova_powervm/tests/virt/powervm/test_vif.py

Change-Id: Iad1197218cb8d8a260c00d55c9addd39ac3a6226
Closes-Bug: #1629556