Create/update vNICs properly on Plug
When deciding whether to "re-treat" or create a VIF, the PlugVifs Task
was only looking for CNAs with the matching MAC address.  With SR-IOV
VNIC, the VIF's MAC lives on the VNIC, not the CNA.

This change set detects whether a VIF is a VNIC; and if so, looks for a
matching VNIC rather than a matching CNA.

This change set also fixes a regression whereby PlugMgmtVif was only
looking in the list of pre-existing CNAs to see whether the management
VIF is already plugged.  We now also keep track of which CNAs we create
during PlugVifs and pass them along.  For efficiency, if we never look
up the existing CNAs, we pass None through, and PlugMgmtVifs looks up
the CNAs afresh.

Change-Id: I892f61d8cd3b07792d0cb16f4aaa643df3ec979c
Closes-Bug: #1619743