Send custom vif plug/unplug event
The main vif plug/unplug flows now push a custom Event through PowerVM
including the MAC address of the vif.  This Event is intended to be
consumed by the PowerVM neutron agents as a lightweight means of knowing
when they should update the corresponding neutron port's status.

The agents must still account for missed events.

Change-Id: I4ba159071f633b2817939de036bca378ead110a2