PlugVifs only provides CNAs
When PlugVifs was invoked with a mix of CNA and VNIC networks, we were
adding VNICs to the list of "CNAs" provided (returned) by the Task.
This was blowing up in PlugMgmtVif (the consumer of that list) because
VNIC doesn't have a vswitch_uri property.

This change set adds an explicit check to ensure the returned list only
includes CNAs.

Change-Id: If1c4ac78d4cc048fcd299dc696b06a010eea18e5
Closes-Bug: #1628955