Fix code for new pep8 rules
This change updates the code to comply with a new pep8 rule to defer the
string concatenation to the logger instead of doing it inline.

It also adds an ignore rule for conf options being outside of nova.conf.
Given that we are an out-of-tree driver, we can not put our conf options
in tree.  This is a temporary work around until the powervm driver is in
tree.

Change-Id: I8fe77e8250a2b59ae842cdacfa48baf66356ee50
Closes-Bug: #1632441