Support segregated function for upload
The original PowerVM Glance code used a model where it would iterate
over the chunks and build a facade on top of that to make it look like a
stream.  We found that starting in Glance v2 that is not incredibly
reliable.

This has led to some bad deploys and reschedules on new hosts.

With this change set, we take advantage of a new feature in pypowervm
that allows us to pass in a function for the upload.  The function
passes in the path of where the image should be streamed to.  We can use
this to allow for the driver to just tell Glance to stream to a file.

Our initial tests shows this to be significantly more reliable (no
failed deploys since).  It also has an ancillary effect where the
download runs about two times faster.

Change-Id: I0fcc630b55290d6bb8fcd82230b58bed81afa4f8
Closes-Bug: #1636321