Better error handling of QP errors
Quick property URIs could sometimes not embed a response.  This is
generally an API issue and changes have gone into pypowervm for it.
However, nova-powervm should also be more tolerant of this.

This change supports better toleration in the event that the error
raised from the REST API does not have a proper response.

Change-Id: I490731a342a3045887a47c1051448e9cd2eabb15
Closes-Bug: 1635385