Refactor events into own class
The eventing model needs to be updated to handle delayed events (and in
future needs more logic to handle resumes on target during LPM).  There
is going to be additional complexity in events.  This change set moves
it out into its own class so as not to keep growing the already long
driver.py file.

Change-Id: I1eadc930d9a55a5fb92f7cd6ffdd37965bb187b6
Partial-Bug: 1617746