Add delay queue for events
Certain events can come in a series.  For instance, if you run a VM
reboot, the hypervisor will emit a SHUTTING_DOWN, NOT_ACTIVATED and
STARTING state.

If these states are just fed to nova raw, due to the async nature it
could come in out of order.  So the ACTIVE event could come in before
the NOT_ACTIVATED (even though that doesn't reflect how it came in to
the host).

This change follows the libvirt event mechanism and builds a delay
queue.  For certain powering off events, they will be delayed before
they are emitted up to nova.  If a corresponding starting event comes in
afterwards, it will void out the shutting down event.

Change-Id: Iefd33b6bf9e3621a9f716488bf43b02c7d3cb0c4
Closes-Bug: 1617746