Clean up of OVS port on source host post migration
After live migration of an instance, as part of the cleanup,
the deletion of port from OVS is not happening on source host.

With this fix, the OVS port is cleaned up as part of the
post live migration on source.

Change-Id: I83ac822c39b4dcc023f71d7f3766ceefe52a8022
Closes-Bug: #1637551
(cherry picked from commit 7c968936006fa51a8a463482412346a8f71f17dd)