New pep8 rules, upper constraints
This change updates the code to comply with a new pep8 rule to defer the
string concatenation to the logger instead of doing it inline.

It also adds an ignore rule for conf options being outside of nova.conf.
Given that we are an out-of-tree driver, we can not put our conf options
in tree.  This is a temporary work around until the powervm driver is in
tree.

And the tox.ini needs to make use of the upper constraints.  Failure to
do so is causing us to hit known issues with newer versions of packages.

And the tox.ini needs to be pulling the nova project from stable/mitaka.

And having done that, older versions of test utils don't have some of
the assertion methods we were using, so those had to be fixed.

Change-Id: I8260acca9de5b21ff6539a509d6b29e9603b9d7f
Closes-Bug: #1632441
Closes-Bug: #1637938