Update the init of the collector to take in a conf
In the Ocata release, there is a new conf variable that is passed
into the collector.  This is generally used by the libvirt collector
at the moment.  It is not needed by the PowerVM collector, but the
__init__ signature needs to be updated to take it in.  This change
brings the PowerVM collector in line with the base requirements from
the core ceilometer project.

Change-Id: Ib5620f6830949f50939d07ba07be900a5451df56