ProvisionRequest does not distinguish event source
This patchset is to support review
https://review.openstack.org/#/c/396467/ of networking_powervm
patchset. ProvisionRequest in networking_powervm is built on
event generated by nova_powervm when a vif is plugged or unplugged.
It needs a way to identify source of such an event. This patchset
provides it. When a plug or unplug event is generated, corresponding
vif_type is also passed into it under details attribute.

This patchset is a cherry-pick from master.
This second patchset removes references to instance variable
introduced using this patchset in master:
https://review.openstack.org/#/c/385144/

Change-Id: I6266ad485b10bdfa21a7b4e66600dfbd6ab1f2fc
Closes-Bug: #1640564