Add tox_install.sh for pypowervm upper-constraints workaround
With the recent addition of pypowervm to upper-constraints, installs
of ceilometer-powervm are now by default enforcing the version set there.
Since the master branch of ceilometer-powervm depends on the develop
version of pypowervm, this causes installs to fail.

This changeset adds a wrapper script (tox_install.sh) that edits
the upper-constraints file and replaces the pypowervm version pin
then sets pypowervm to be installed from source instead.

Change-Id: Ib60ee84ecb168991439ac5d097ba07a62d2329a7