Object upload in swift occasionally fails with 'TypeError'.
Data upload to swift store occasionally fails due to invalid
keystone token. In which case, the retry operation uses the same
StringIO object for the source. As it was read in the
first attempt, it will end up with a data of length 0 byte causing
this issue. Moving it inside the retry block, so it constructs a
new one on each retry.

Change-Id: Ic335815a25fb18b8b547add4e5decd531ca1d93d
Closes-Bug: #1642489