First pass at imagecache
Imagecache checks for LV with names like i_* and
compares that list to a list of currently used images. It will
then remove all images not currently being used.

Localdisk will check if there is a cached image available, if there
is, use it. Otherwise create a cached image to use. Then use it.

Change-Id: Ibc60d633153ca0cfa2b1c8c3e0cbdb80a1546409
Implements: bp image-cache-powervm