Migration scenario, SR-IOV vif contains source physical ports
This is not a bug, this is related to future work to support migration for
SR-IOV VMs. In migration scenarios, SR-IOV plug mechanism receives vif and
it proceeds to create vNIC. This fails due to physical ports mismatch. As
there are no physical ports matching with location codes in incoming vif,
no backing device could be created.

This is due to the fact that rebuild logic in nova compute, retrieves vif
details from nova instance and feeds it to plug mechanism. Though neutron
SR-IOV mechanism driver provides new and valid physical ports corresponding
to destination host, it is ignored.

A fix is needed in SR-IOV plug mechanism to retrieve appropriate SR_IOV
physical ports that corresponds to related physical network (from related
neutron network) and consume.

There are two parts to this fix. One is related to physical_network. If
there are no physical network in incoming vif detail, it is retreived using
network id. Refer to
https://bugs.launchpad.net/networking-powervm/+bug/1651129
for a fix on networking-powervm to avoid this moving forward.

Second part is to retrieve physical ports from the environment using
physical_network against port label attribute.

Change-Id: Ic511314095472da21c3a5588d77955c939b7e0f1