Switch to synchronous power on
The power on action is a job.  Previously the powervm driver would start
the job but not wait for it to complete.  This caused issues with
automation that sat on top of the OpenStack that the driver was part of.
The issue occurred because the power on job would be issued, but then
OpenStack would 'sync the power state' before the job completed.  This
caused an inconsistency between the states in OpenStack and the VM
itself.

This flew under the radar for a while because the power on job is quite
quick.  But this has started popping up on novalinks that have done
thousands of deploys or are at a bit higher deploy load.

Change-Id: I7ebef04bed4060bd8df9fb238b5965cd685b2d03
Closes-Bug: 1654262