Add vopt removal params to the power on job in spawn
The PowerVM platform is introducing two new parameters to the power
on job for removal of virtual optical media after X amount of
time.

For initialization of VMs, nova-powervm makes use of the
config drive attached to the LPAR on spawn which currently has
to be manually removed. This change set adds the config option to
leverage the new function in the platform to remove the vopt media
from the VM X minutes (default 60) after spawn.

Change-Id: I572d424d9cd5ebbac20118d88cdb2f18c3b16521