Update File i/o to support live-migration and use slot map
Add pre_live_migration_on_destination to fileio to support live
migration.

This update modifies the file i/o cinder connector to support
pulling/saving the slot. The slot map indicates the position in
the SCSI bus for a given disk. Also allow fileio type to be in
the recreate map.

Change-Id: Idcef2198d221e7faa832a59cf7e57297d1c23388
Closes-Bug: 1662995