Synchronize power ops
When separate threads are issuing power on/off requests on a VM
(possibly through spawns, deletes, power_on, power_off, reboot, etc.),
timing issues can arise when threads are retrieving and acting on the VM
state. This change set synchronizes power on/off operations per
instance to avoid this.

A side-effect of this change is that the PowerOn Task no longer requires
the lpar_wrap from prior Tasks.

Change-Id: Ia21a3f98f51f338115d3728702881bc747f67b9f
Closes-Bug: #1664720