Get flavor from instance
Flavor is now available directly as instance.flavor, and is no longer
included in the argument spec for spawn.

This change set removes the flavor kwarg from spawn, and all use of
Flavor.get_by_id.  In so doing, the following flavor and related
parameters are removed from internal APIs:

nova_powervm.virt.powervm.
  tasks.vm.Create: flavor
  tasks.vm.Resize: flavor
  vm.crt_lpar: flavor
  vm.update: flavor
  vm.VMBuilder.lpar_builder and _format_flavor: flavor
  tasks.storage.CreateDiskForImg: disk_size
  disk.driver.DiskAdapter.create_disk_from_image,
      _create_disk_from_image, and subclass impls thereof: disk_size

Some minor cleanup is also included in this change.

Change-Id: I7e619cf1ab9b6d428e177d0d20e805f13539b3b2