Config Option for Endpoint Type for Swift
The Swift client will use the Keystone Service Catalog to lookup the
URL to use when connecting to the various Swift Services.  By default
it will use the 'publicURL' endpoint/interface type for which URL to
use.  However, depending on the configuration of the environment, it
may be desired to use another endpoint/interface type, such as
"internalURL".

This change will add a configuration option, similar to the others
provided in nova-powervm for Swift client options, to be able to
specify this endpoint/interface type in a 'swift_endpoint_type'
property that will map to the os_endpoint_type option for Swift.

Change-Id: Id6421cc167ef92d31a7ebcb2a8ffb43e6bc72412