Check for presence of volume data before doing cleanup
In case of a failure in pre live migration phase the volume data
will not be present. In such case when cleanup_volume is invoked
it results in an exception. The exception is harmless and does
not affect the migration flow but it can be avoided by adding a
simple check for presence of volume data.

Change-Id: I7dcb8b63140039bbf8025d2bcaf1add89f9ece7f
Closes-bug: 1678132