Update host cpu util calculation to consider idle proc cycles
The Lpar utilization used to consider donatedProcCycles. As
per discussion with hypervisor team, donatedProcCycles are
not guaranteed to be utilised by any other Lpar on the host. So,
from host perspective, it might get utilised by another lpar as uncapped
or it might not get utilised. Since we are already including uncapped
proc cycles in calculation, we need to ignore donated. Also, the
formula needs to take into account idle proc cycles on lpar.

Change-Id: Ia26d16796be21320eb64f2be15f958c1d0b0d174
Closes-Bug: 1686422