Allow vios to be filtered for rebuild
During rebuild process all the active VIOS are considered as valid.
In case of some powervm env a VIOS might not own any IO and is not
a valid candiate for rebuild.

Adding a vios_uuids method to define the valid vios for a given
volume driver. This property can be used further in all the
connect and disconnect operations as well. That will be done as
part of a future changeset.

Change-Id: I936c12a8b1fff431fcf3098aca57cedcdbf32b0b