In Tree Backport: Exception handling #1
This backports a few changes regarding exception handling from the
in-tree driver. It ensures that LOG.exception is passed a meaningful
error message whenever called and is passed the instance when possible.
It also removes the use of the raise keyword on its own, replacing it
with the excutils.save_and_reraise_exception() function from oslo_utils.
This is the first of a series of changes to the OOT driver exception
handling.

Change-Id: I7e1039360399d6861ab4258b684324fd3d880f1d