Add snapshot capability in DiskAdapter
Add a 'snapshot' capability in the DiskAdapter superclass, and check it
from the virt driver's snapshot method.  This allows us to generate a
prettier error when snapshot is attempted with a disk driver that
doesn't support it.

The snapshot capability is True for the SSP driver.

The snapshot capability is True for the localdisk driver if the
management partition is acting as the VIOS; but False for traditional
VIOS.

Change-Id: Ide8c495a4293050f9844e16132b7b751ec50cbf1
Closes-Bug: #1661395