Change NVRAM manager and Swift APIs to accept UUID
Currently the NVRAM manager store API takes the instance object as
the paramater, which by itself is just fine. However, the event handler
(virt/powervm/event.py) has to grab the entire Instance object just for
the store method when responding to NVRAM events.

This change makes it so that the NvramManager and Swift APIs take
the instance uuid OR instance object as the parameter. A cache is put
in the event listener to maintain a mapping of the PVM UUID to the
instance object UUID so that the instance object still does not have
to be retrieved on every event.

Change-Id: Id3513cf9a7d905f772f1eb3576a3e79a6fb57eae