Use CONF.host for powervm nodename
The get_available_nodes method of the powervm virt driver was
returning a string based on the machine type, model, and serial
number of the host. This was inconsistent with other drivers. It also
required a working NovaLink connection, which cannot be guaranteed.
This change replaces the MTMS string with CONF.host.

The same change was made to the in-tree driver with nova commits
f2e8244e169ed2ecb8e06e76de1755033be87f21 and
de5b7d8f257ba89d5f3dab805f8e49ffc6eda4aa

Change-Id: Id41c4576c32e7e8cf6b2c3c2020303f87ecaf3d3