Support PPT ratio capable to PPT ratio incapable host rebuild
We don't want to block rebuilds from PPT ratio supported hosts to
hosts that don't support setting the PPT ratio. If we're rebuilding
on a host that doesn't support setting the PPT ratio, ignore
setting the value in the LPARBuilder at all so that the validation
logic doesn't block it.

Change-Id: Iad44f4941bafcb48a1f0c98ff38a24bae39c9eb6