Add extend volume support for fileio
When cinder resizes a volume, it sends an event to nova. This event
calls into the compute driver's "extend_volume" method. Calling the
pypowervm rescan_vstor is all that is needed for our driver.

Only kernel backstores are currently supported by the rescan_vstor
method.

Not all vios are guaranteed to have the storage mapped. Log an error
if the virtual disk was not found, but throw an error if no virtual
disks were found on any of the vios.

Switching the backstore from USER_QCOW to FILE_IO enables this change.
This adds support to extend size of a volume while it is attached to
a client VM.

Change-Id: I0f6f80df6ad0d87dcccc560509c4e31992e4c0ec