Persist existing LPAR wrapper attributes in DefaultStandardize on resize
The DefaultStandardize object in pypowervm defaults the proc units
factor, max virtio slots, uncapped weight, spp, availability
priority, srr, proc compat, and lpar metric values. nova-powervm
does not pass in these parameters when resizing.

This results in these attributes being set to the default defined
by the DefaultStandardize object for every resize operation. This
change passes those parameters into the DefaultStandardize object.

Note that if the attributes were passed in as part of the flavor
then the flavor's values would be set later in the DefaultStandardize
object by the LPARBuilder.

Closes-Bug: #1722923
Change-Id: I0e9ad0dbbc35658e096636a454ad493492025420