stop setting extra capabilities
Only one of the three current disk driver capabilities is also
supposed to be a capability of the virt driver, yet all of them
were being copied onto the virt driver. This corrects that so
that only has_imagecache appears in both places.

It also fixes how we test capabilities, which was a) incomplete and
b) only passing coincidentally for has_imagecache (i.e., not testing
what it intended to be testing). This required making capabilities
an instance property instead of a class property, so that when one
test sets a capability that doesn't bleed over into other tests.

Change-Id: I66928342c9dc8c2735b739d1b96f0d5b187b9e8c