update translations for queens
Update translations for the Queens release.

Note: this is being cherry-picked from stable/queens to master
instead of the other way around. The translations were actually done
based on queens, so that is where they were originally merged. In
hindsight, they should also be merged in master so that it has the
latest translations as well, so we should have followed the normal
process of merging in master first, but too late for that now.

Change-Id: I0622d237a53114929d56cfb46bbb7d60887099d6
(cherry picked from commit e62fae699d47feacd7372573948722cd0036996c)