Fix boot volume connectivity type discovery
Per the referenced bug, there was a code path whereby the boot volume
connectivity type could be unset, resulting in an exception.  In fixing
this, I discovered that we were looking for `driver_volume_type` in the
wrong place: it should be in connection_info and we were looking for it
under connection_info['data'].  I preserved the bogus code path in case
it's actually somehow legit, but added a path to the correct location.

We had no direct test for this stuff, so I added that too.

Change-Id: I78e55b3d8096d4013afd0abe1e42c32dc9697871
Closes-Bug: #1769648