Fix requirements and lower constraints
We were using the lower-constraints.txt file from nova, but this
was masking problems. The check job doesn't seem to work properly
when the lower-constraints.txt file is in a different project, not
running at least some of its tests. This was seen when copying the
same file into nova-powervm cause the job to start complaining.

This copies nova's lower-constraints.txt file into nova-powervm
and then addresses the aforementioned issues by:

1) removing amqp from requirements.txt. We do not appear to
really require this.
2) adding bashate to lower-constraints.txt. Nova does not require
this but we do.
3) adding python-swiftclient to lower-constraints.txt. Nova does not
require this but we do. Also moved python-swiftclient from
test-requirements.txt to requirements.txt since our usage is not
limited to tests.
4) removing python-subunit from test-requirements.txt. We do not
appear to really require this.

Change-Id: I100181821ed4ffe622ed5ebeff403093e9cc32b8