tox: Don't create '.pyc' files
Our tox.ini was trying to use find but failing because it was not
added to whitelist_externals. Rather than correct that, we can
remove it entirely by adjusting tox to no longer create '.pyc'
files, so we no longer have to delete them.

Nova previously did this [1] to improve performance, but then had
to revert it [2] because of a tox bug. That bug has since been
fixed [3], so we should be ok to do this.

[1] 0d6d1616b9a05f1c3e1e34cb235fee327196d423
[2] 233160644f4f0a17177fffe8a1e6db1341e61ad4
[3] https://review.openstack.org/#/c/582392/1/tox.ini@28

Change-Id: I11cbf81160ba2eeefcce19ed25930c39e17449b7