More py3.x support
The Stein supported runtimes proposal [1] has chosen py36 over py35
due to the communities move to testing on Ubuntu 18.04, where py36
is the default runtime. We update tox.ini and setup.cfg accordingly.

We also add a py37 option to the tox.ini to facilitate future-looking
tests per the resolution on keeping up with Python 3 releases [2].

The py35 option will eventually need to be removed from tox.ini, but
this keeps it for now in order to allow time for folks to transition.

[1] https://review.openstack.org/611080
[2] https://review.openstack.org/613145

Change-Id: Id77fbd969bee5a6f3636be13362a2b89cb5888e8