Add the device_id parameter to discover_hdisk
LUA recovery, the job that tells the VIOS to discover a specific
hdisk, may find multiple disks at the same ITL. Without the device ID
LUA recovery will return an error when this happens.

For most volume backed storages the device ID is the base 64 encoded
value of the pg83 identifier. This change enables the passing of the
deviceID to discover_hdisk to avoid the multiple hdisks at the same
ITL error. It also bumps the pypowervm version for the device_id
support in discover_hdisk.

Change-Id: I86c273bf1d1c5753ce68ab51f6922c323d720cb2